Several changes to make dependencies less of a problem.

+++ libs/core/Execution_Object.h
@@ -21,7 +21,9 @@ namespace polaris
 		
 		Base_Object(size_t component_id,int uuid): _component_id(component_id),_uuid(uuid){}
 
-		size_t _component_id;
+		//virtual void Allocate_Vtable(){}
+
+		unsigned int _component_id;
 		int _uuid;
 	};
